requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.5
kafka_python==2.0.2
python-decouple==3.8
Requests==2.31.0
